Make isValidEditNote a reusable function
There's no real reason I can see why we wouldn't want to check for
invalid edit notes already before submission in all React forms,
but we only do it in the release editor right now (all others are
checked by FormHandler).

This makes the checking function reusable so that we can later
use it in any relevant forms.
